# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DocumentType": {
				"type": "string",
				"enum": [
					"logo",
					"stamp",
					"signature",
					"avatar",
					"survey",
					"other",
					"plan",
					"pdf",
					"employee",
					"medical-fitness-certificate",
					"work-authorization-document",
					"personal-protective-equipment-inspection-report",
					"annual-subcontractor-framework-agreement",
					"annual-general-occupational-safety-training",
					"data-processing-declaration",
					"tig",
					"invoice",
					"before",
					"during",
					"completion",
					"inspection",
					"contract"
				]
			},
			"DocumentOwnerType": {
				"type": "string",
				"enum": [
					"user",
					"tenant",
					"contact",
					"company",
					"tender",
					"project",
					"milestone",
					"report",
					null
				],
				"nullable": true
			},
			"DocumentProperties": {
				"properties": {
					"startOfValidity": {
						"type": "string"
					},
					"endOfValidity": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Document": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"$ref": "#/components/schemas/DocumentType"
					},
					"ownerId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"ownerType": {
						"$ref": "#/components/schemas/DocumentOwnerType"
					},
					"mimeType": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"properties": {
						"$ref": "#/components/schemas/DocumentProperties"
					},
					"companyId": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						],
						"nullable": true
					},
					"approved": {
						"type": "boolean"
					},
					"stored": {
						"properties": {
							"thumbnail": {
								"type": "string"
							},
							"resized": {
								"type": "string"
							},
							"original": {
								"type": "string"
							}
						},
						"required": [
							"original"
						],
						"type": "object"
					}
				},
				"required": [
					"type",
					"mimeType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_Document_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/DocumentType"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"ownerType": {
						"$ref": "#/components/schemas/DocumentOwnerType"
					},
					"mimeType": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"properties": {
						"$ref": "#/components/schemas/DocumentProperties"
					},
					"companyId": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"approved": {
						"type": "boolean"
					},
					"stored": {
						"properties": {
							"thumbnail": {
								"type": "string"
							},
							"resized": {
								"type": "string"
							},
							"original": {
								"type": "string"
							}
						},
						"required": [
							"original"
						],
						"type": "object"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"LOCATION_STATUS": {
				"enum": [
					"active",
					"inactive"
				],
				"type": "string"
			},
			"Partial_Location_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"taxNumber": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"region": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/LOCATION_STATUS"
					},
					"notes": {
						"type": "string"
					},
					"tenantId": {
						"type": "number",
						"format": "double"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"CONTACT_STATUS": {
				"enum": [
					"active",
					"inactive"
				],
				"type": "string"
			},
			"Partial_Contact_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/CONTACT_STATUS"
					},
					"notes": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"tenantId": {
						"type": "number",
						"format": "double"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"COMPANY_STATUS": {
				"enum": [
					"active",
					"inactive"
				],
				"type": "string"
			},
			"COMPANY_TYPE": {
				"enum": [
					"customer",
					"contractor",
					"supplier"
				],
				"type": "string"
			},
			"Partial_Company_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"region": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/COMPANY_STATUS"
					},
					"type": {
						"$ref": "#/components/schemas/COMPANY_TYPE"
					},
					"default": {
						"type": "boolean"
					},
					"taxNumber": {
						"type": "string"
					},
					"registrationNumber": {
						"type": "string"
					},
					"bankAccount": {
						"type": "string"
					},
					"prefix": {
						"type": "string"
					},
					"offerNum": {
						"type": "string"
					},
					"tenantId": {
						"type": "number",
						"format": "double"
					},
					"notes": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double"
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/Document"
						},
						"type": "array"
					},
					"contacts": {
						"items": {
							"$ref": "#/components/schemas/Contact"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"Omit_Company.id-or-createdOn-or-updatedBy-or-updatedOn_": {
				"$ref": "#/components/schemas/Pick_Company.Exclude_keyofCompany.id-or-createdOn-or-updatedBy-or-updatedOn__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateCompanyProperties": {
				"$ref": "#/components/schemas/Omit_Company.id-or-createdOn-or-updatedBy-or-updatedOn_"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Vilkrig x Alpinnovacio CRM API",
		"version": "1.0.0",
		"license": {
			"name": "UNLICENSED"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/{id}/documents": {
			"put": {
				"operationId": "UploadDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Updates the profile picture of a user by ID.\nSecured with JWT token and requires \"User:Update\" or \"Me:*\" permission.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwtToken": [
							"User:Update",
							"Me:*",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the user to update the profile picture.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/DocumentType"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			}
		},
...
		"/tenants/{id}/documents": {
			"put": {
				"operationId": "UploadDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"uploaded": {
											"type": "boolean"
										}
									},
									"required": [
										"uploaded"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Uploads or updates a logo for a specific tenant.The endpoint requires authentication\nand the \"Tenant:Update\" permission.The logo is associated with the tenant based on their ID.",
				"tags": [
					"Tenant"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the tenant to update the logo for.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			}
		},
...
		"/projects/{id}/documents": {
			"get": {
				"operationId": "GetDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_Document_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves detailed information about a specific tender documents by their ID.\nThis endpoint is protected by JWT authentication, requiring \"Project:GetDocuments\" permission.",
				"tags": [
					"Project"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Project:GetDocuments"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UploadDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Upload documents for the specified tender by their ID.\nThis endpoint is protected by JWT authentication, requiring \"Project:UpdateDocuments\" permission.",
				"tags": [
					"Project"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Project:UpdateDocuments"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/DocumentType"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			}
		},
...
		"/companies": {
			"get": {
				"operationId": "GetCompanies",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_Company_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of companies filtered by type. This endpoint is secured with JWT authentication and requires\nspecific permissions. It supports filtering companies by their type, either CONTRACTOR or CUSTOMER.\n\nAuthorization tokens and user permissions are validated before processing the request. The user's tenant information\nis utilized to fetch relevant company data.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:List"
						]
					}
				],
				"parameters": [
					{
						"description": "The type of companies to filter by. Accepts either 'CONTRACTOR' or 'CUSTOMER'.",
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {}
					}
				]
			},
			"post": {
				"operationId": "CreateCompany",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Company_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Creates a new company with the provided details. This endpoint requires authentication and specific permissions.\nIt utilizes the user's tenant and ID from the JWT token for creating the company in the correct context.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The properties required to create a new company.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCompanyProperties",
								"description": "The properties required to create a new company."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCompanies",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes multiple companies based on their IDs. This is a secured endpoint that validates the user's authentication\nand permissions. The operation uses the user's tenant information for authorization.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Delete"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Contains the list of company IDs to delete.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ids": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"ids"
								],
								"type": "object",
								"description": "Contains the list of company IDs to delete."
							}
						}
					}
				}
			}
		},
		"/companies/{id}": {
			"get": {
				"operationId": "GetCompany",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Company_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Fetches detailed information about a specific company identified by its ID. This method ensures that only\nauthenticated users with appropriate permissions can access company details. The endpoint uses JWT for security.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Get"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the company to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCompany",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Company_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Updates an existing company's details identified by its ID. This endpoint checks for valid authentication\nand the necessary permissions before allowing updates to proceed. The user's tenant and ID are used for\nauthorization and ensuring data consistency.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the company to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The new details for the company. Partial updates are supported.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Company_",
								"description": "The new details for the company. Partial updates are supported."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCompany",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes a single company identified by its ID. Like other endpoints, this operation requires authentication\nand appropriate permissions. The user's tenant information is critical for ensuring secure data manipulation.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Delete"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the company to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/companies/{id}/documents": {
			"get": {
				"operationId": "GetDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Fetches detailed information about a specific company identified by its ID. This method ensures that only\nauthenticated users with appropriate permissions can access company details. The endpoint uses JWT for security.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Get"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the company to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UploadDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Uploads or updates a company's logo. This action is protected and requires authentication and specific permissions.\nThe logo is linked to the company's profile based on the company ID.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The company ID to associate the logo with.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/DocumentType"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			}
		},
		"/companies/{id}/contacts": {
			"post": {
				"operationId": "AddContacts",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Adds multiple contacts to a company specified by the company ID. This operation is secured with JWT and requires\n\"Company:Update\" permission, ensuring that only authorized users can add contacts to the company.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the company to which contacts will be added.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "An array of contacts, represented as partial objects, to be added to the company.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/Partial_Contact_"
								},
								"type": "array",
								"description": "An array of contacts, represented as partial objects, to be added to the company."
							}
						}
					}
				}
			}
		},
		"/companies/{id}/locations": {
			"post": {
				"operationId": "AddLocations",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Adds multiple locations to a company specified by the company ID. This method is secured with JWT and requires\n\"Company:Update\" permission, allowing only authorized users to add locations to the company.",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Company:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the company to which locations will be added.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "An array of location objects, represented as partial objects, to be added to the company.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/Partial_Location_"
								},
								"type": "array",
								"description": "An array of location objects, represented as partial objects, to be added to the company."
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}